backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
aws-cdk-lib==2.111.0
aws-cdk.asset-awscli-v1==2.2.201
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
aws-cdk.aws-lambda-python-alpha==2.111.0a0
boto3==1.29.6
botocore==1.32.7
cattrs==23.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
constructs==10.3.0
coverage==7.3.2
cryptography==41.0.7
distro==1.8.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-resources==6.1.1
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==1.0.1
jsii==1.92.0
MarkupSafe==2.1.3
moto==4.2.10
openai==1.3.6
packaging==23.2
pluggy==1.3.0
publication==0.0.3
py==1.11.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-mock==1.11.0
responses==0.24.1
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
toml==0.10.2
tomli==2.0.1
tqdm==4.66.1
typeguard==2.13.3
typing_extensions==4.8.0
urllib3==1.26.18
Werkzeug==3.0.1
xmltodict==0.13.0
zipp==3.17.0
